SpaceX
The project SpaceX Falcon 9 first stage stage landing successor predictor is a
machine learning project that aims to predict the likelihood of a successful
landing of the Falcon 9 first stage. The project will use a dataset of historical
launch data to train a machine learning algorithm. The algorithm will then be used
to make predictions about the likelihood of a successful landing for future
launches.